refactor(auth): modularize bcrypt and jwt, use kebab-case on file names
