+++ auth/lib/bcrypt.js
@@ -1,0 +1,5 @@
+const bcrypt = require('bcrypt');
+
+exports.hash = password => (bcrypt.hash(password, 8));
+
+exports.compare = (password, hash) => (bcrypt.compare(password, hash));
