+++ auth/lib/error.js
